pyproject.toml
[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.12"
python-dotenv = "^1.1.0"
sqlalchemy = "^2.0.40"
sqlmodel = "^0.0.24"
asyncpg = "^0.30.0"
uvicorn = "^0.34.0"
greenlet = "^3.1.1"
google-api-python-client = "^2.155.0"
google-auth = "^2.35.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
psycopg2-binary = "^2.9.10"
alembic = "^1.13.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
